+++ src/checks.c
@@ -2,6 +2,7 @@
 #include <string.h>
 #include "checks.h"
 #include "is_integerish.h"
+#include "is_sorted.h"
 #include "any_missing.h"
 #include "any_infinite.h"
 #include "all_missing.h"
@@ -118,7 +119,7 @@ static Rboolean check_bounds(SEXP x, SEXP lower, SEXP upper) {
 }
 
 
-Rboolean check_posix_bounds(SEXP x, SEXP lower, SEXP upper) {
+static Rboolean check_posix_bounds(SEXP x, SEXP lower, SEXP upper) {
     if (isNull(lower) && isNull(upper))
         return TRUE;
 
@@ -354,57 +355,9 @@ static inline Rboolean is_scalar_na(SEXP x) {
     return FALSE;
 }
 
-static Rboolean is_sorted_integer(SEXP x) {
-    R_xlen_t i = 0;
-    const int * const xi = INTEGER(x);
-    const R_xlen_t n = xlength(x);
-    while(xi[i] == NA_INTEGER) {
-        i++;
-        if (i == n)
-            return TRUE;
-    }
-
-    for (R_xlen_t j = i + 1; j < n; j++) {
-        if (xi[j] != NA_INTEGER) {
-            if (xi[i] > xi[j])
-                return FALSE;
-            i = j;
-        }
-    }
-    return TRUE;
-}
-
-
-static Rboolean is_sorted_double(SEXP x) {
-    R_xlen_t i = 0;
-    const double * const xr = REAL(x);
-    const R_xlen_t n = xlength(x);
-    while(xr[i] == NA_REAL) {
-        i++;
-        if (i == n)
-            return TRUE;
-    }
-
-    for (R_xlen_t j = i + 1; j < n; j++) {
-        if (xr[j] != NA_REAL) {
-            if (xr[i] > xr[j])
-                return FALSE;
-            i = j;
-        }
-    }
-    return TRUE;
-}
-
-
 static Rboolean check_vector_sorted(SEXP x, SEXP sorted) {
     if (asFlag(sorted, "sorted") && xlength(x) > 1) {
-        Rboolean ok;
-        switch(TYPEOF(x)) {
-            case INTSXP: ok = is_sorted_integer(x); break;
-            case REALSXP: ok = is_sorted_double(x); break;
-            default: error("Checking for sorted vector only possible for integer and double");
-        }
-        if (!ok)
+        if (!isSorted(x))
             return message("Must be sorted");
     }
     return TRUE;
@@ -413,7 +366,7 @@ static Rboolean check_vector_sorted(SEXP x, SEXP sorted) {
 /*********************************************************************************************************************/
 /* Exported check functions                                                                                          */
 /*********************************************************************************************************************/
-SEXP attribute_hidden c_check_character(SEXP x, SEXP min_chars, SEXP any_missing, SEXP all_missing, SEXP len, SEXP min_len, SEXP max_len, SEXP unique, SEXP names, SEXP null_ok) {
+SEXP attribute_hidden c_check_character(SEXP x, SEXP min_chars, SEXP any_missing, SEXP all_missing, SEXP len, SEXP min_len, SEXP max_len, SEXP unique, SEXP sorted, SEXP names, SEXP null_ok) {
     HANDLE_TYPE_NULL(isString(x) || all_missing_atomic(x), "character", null_ok);
     ASSERT_TRUE(check_vector_len(x, len, min_len, max_len));
     ASSERT_TRUE(check_vector_names(x, names));
@@ -424,6 +377,7 @@ SEXP attribute_hidden c_check_character(SEXP x, SEXP min_chars, SEXP any_missing
             return result("All elements must have at least %i characters", n);
     }
     ASSERT_TRUE(check_vector_unique(x, unique));
+    ASSERT_TRUE(check_vector_sorted(x, sorted));
     return ScalarLogical(TRUE);
 }