+++ src/is_sorted.h
@@ -1,0 +1,11 @@
+#ifndef CHECKMATE_IS_SORTED_H_
+#define CHECKMATE_IS_SORTED_H_
+
+#define USE_RINTERNALS
+#include <R.h>
+#include <Rinternals.h>
+#include <Rversion.h>
+
+Rboolean isSorted(SEXP x);
+
+#endif